fix: implement hoc with handleClick
