+++ webpack.config.js
@@ -18,6 +18,7 @@ module.exports = {
       '@store': path.resolve(__dirname, 'src/store'),
       '@utils': path.resolve(__dirname, 'src/utils'),
       '@styles': path.resolve(__dirname, 'src/styles'),
+      '@hoc': path.resolve(__dirname, 'src/hoc'),
     },
   },
   module: {
